ESP32_WROOM_32_&_A9G_chip_configuration
The solution presents a software connection between the ESP32_WROOM_32 chip and the A9G GSM/GPRS/GPS module. Could be the first of its kind on the internet as other online references still show a difficulty in making the A9G configuration with other controller modules. Some parts are still under development.